requirements.txt
plyara==2.2.8
yara-python==4.5.4
